null

+++ tmbeans/src/main/java/com/modcreater/tmbeans/pojo/DiscountCoupon.java
@@ -13,7 +13,7 @@ public class DiscountCoupon {
   private Long id;
   private String couponName;
   private Long couponType;
-  private Long couponMoney;
+  private Double couponMoney;
   private Long starTime;
   private Long entTime;
   private Date createDate;
